+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# 0.0.1
+# 0.0.2
+
+- Change `bodyParser` to parse body in `put`, `post`, `patch` and `delete` http methods.
+- Add default error responce 501 if route for request isn't provided.
+
+## 0.0.1
 
 - Initial version.
 - Creates core of the package. See README.